[#IOCIT-34] Restore origina pn-client.ts
